requirements.txt
gradio
llama-index
llama-index-embeddings-ollama
llama-index-embeddings-huggingface
llama-index-llms-ollama
pymupdf
python-docx